refactored baseline
